# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Adylic Test Api",
        "description": "Adylic Test Sport Api",
        "contact": {
            "email": "luisenesi@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/games/{gameid}/update_score": {
            "patch": {
                "tags": [
                    "game"
                ],
                "description": "Update Game Score",
                "operationId": "updateScore",
                "parameters": [
                    {
                        "name": "gameid",
                        "in": "path",
                        "description": "Id of Game",
                        "required": true,
                        "explode": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "home_score",
                        "in": "query",
                        "description": "Home Score",
                        "required": true,
                        "explode": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "visit_score",
                        "in": "query",
                        "description": "Visit Score",
                        "required": true,
                        "explode": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found"
                    },
                    "200": {
                        "description": "Game"
                    }
                }
            }
        }
    },
...
}